updated bill items order
