+++ src/components/CustomTable.js
@@ -20,6 +20,7 @@ import { useHistory, useRouteMatch } from "react-router";
 import { Search } from "@mui/icons-material";
 import { useEffect, useState } from "react";
 import { CSVLink } from "react-csv";
+import { getComparator, stableSort } from "../utils/sorting";
 
 const SearchContainer = styled("div")(({ theme }) => ({
   display: "flex",
@@ -40,34 +41,6 @@ const StyledInputBase = styled(InputBase)(({ theme }) => ({
   color: "black",
   marginLeft: theme.spacing(1),
 }));
-
-function descendingComparator(a, b, orderBy) {
-  if (b[orderBy] < a[orderBy]) {
-    return -1;
-  }
-  if (b[orderBy] > a[orderBy]) {
-    return 1;
-  }
-  return 0;
-}
-
-function getComparator(order, orderBy) {
-  return order === "desc"
-    ? (a, b) => descendingComparator(a, b, orderBy)
-    : (a, b) => -descendingComparator(a, b, orderBy);
-}
-
-function stableSort(array, comparator) {
-  const stabilizedThis = array.map((el, index) => [el, index]);
-  stabilizedThis.sort((a, b) => {
-    const order = comparator(a[0], b[0]);
-    if (order !== 0) {
-      return order;
-    }
-    return a[1] - b[1];
-  });
-  return stabilizedThis.map((el) => el[0]);
-}
 
 function EnhancedTableHead(props) {
   const {
